feat/60 homesta

+++ src/main/java/com/homeis/board/controller/BoardController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -16,6 +17,7 @@ import com.homeis.board.dto.BoardPaginationResponse;
 import com.homeis.board.dto.Comment;
 import com.homeis.board.dto.Likes;
 import com.homeis.board.model.service.BoardService;
+import com.homeis.util.JWTUtil;
 
 import lombok.RequiredArgsConstructor;
 
@@ -24,6 +26,7 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/board")
 public class BoardController {
 	private final BoardService boardService;
+	private final JWTUtil jwtUtil;
 	
 	@GetMapping("/list")
 	public ResponseEntity<BoardPaginationResponse> list(
@@ -37,8 +40,14 @@ public class BoardController {
 	}
 	
 	@GetMapping("/detail/{id}")
-	public ResponseEntity<?> detail(@PathVariable("id") int id) {
-		Board board = boardService.findById(id);
+	public ResponseEntity<?> detail(@PathVariable("id") int id, @RequestHeader(value = "Authorization", required = false) String tokenHeader) {
+		String tokenId = null;
+		
+		if(tokenHeader != null && tokenHeader.startsWith("Bearer ")) {
+			tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		}
+		
+		Board board = boardService.findById(id, tokenId);
 		
 		if(board == null) return ResponseEntity.status(404).body("요청하신 글을 찾을 수 없습니다.");
 		
@@ -46,7 +55,11 @@ public class BoardController {
 	}
 	
 	@PostMapping("/insert")
-	public ResponseEntity<?> regist(@RequestBody Board board) {
+	public ResponseEntity<?> regist(@RequestBody Board board, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+
+		if(!board.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.insertBoard(board);
 		
 		if(result == 0) return ResponseEntity.notFound().build();
@@ -55,16 +68,24 @@ public class BoardController {
 	}
 	
 	@PutMapping("/update")
-	public ResponseEntity<?> update(@RequestBody Board board) {
-		int result = boardService.updateBoard(board);
+	public ResponseEntity<?> update(@RequestBody Board board, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+
+		if(!board.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
 		
+		int result = boardService.updateBoard(board);
+
 		if(result == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
 		
 		return ResponseEntity.ok().build();
 	}
 	
 	@PutMapping("/delete")
-	public ResponseEntity<?> delete(@RequestBody Board board) {
+	public ResponseEntity<?> delete(@RequestBody Board board,@RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!board.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.deleteBoard(board);
 		
 		if(result == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
@@ -73,7 +94,11 @@ public class BoardController {
 	}
 	
 	@PostMapping("/like")
-	public ResponseEntity<?> likeRegist(@RequestBody Likes like) {
+	public ResponseEntity<?> likeRegist(@RequestBody Likes like, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!like.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.insertLike(like);
 		
 		if(result == 0) return ResponseEntity.notFound().build();
@@ -82,7 +107,11 @@ public class BoardController {
 	}
 	
 	@PutMapping("/like")
-	public ResponseEntity<?> likeDelete(@RequestBody Likes like) {
+	public ResponseEntity<?> likeDelete(@RequestBody Likes like, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!like.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.deleteLike(like);
 		
 		if(result == 0) return ResponseEntity.notFound().build();
@@ -91,7 +120,11 @@ public class BoardController {
 	}
 	
 	@PostMapping("/insert-comment")
-	public ResponseEntity<?> commentRegist(@RequestBody Comment comment) {
+	public ResponseEntity<?> commentRegist(@RequestBody Comment comment, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!comment.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.insertComment(comment);
 		
 		if(result == 0) return ResponseEntity.notFound().build();
@@ -100,7 +133,11 @@ public class BoardController {
 	}
 	 
 	@PutMapping("/update-comment")
-	public ResponseEntity<?> commentupdate(@RequestBody Comment comment) {
+	public ResponseEntity<?> commentupdate(@RequestBody Comment comment, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!comment.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.updateComment(comment);
 		
 		if(result == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
@@ -109,7 +146,11 @@ public class BoardController {
 	}
 	
 	@PutMapping("/delete-comment")
-	public ResponseEntity<?> commentDelete(@RequestBody Comment comment) {
+	public ResponseEntity<?> commentDelete(@RequestBody Comment comment, @RequestHeader("Authorization") String tokenHeader) {
+		String tokenId = jwtUtil.getIdFromToken(tokenHeader.substring(7));
+		
+		if(!comment.getUserId().equals(tokenId))return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("인증 실패");
+		
 		int result = boardService.deleteComment(comment);
 		
 		if(result == 0) return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("권한 없음. 잘못된 접근.");
